chore: more minimal
