+++ packages/zoe/src/contracts/automaticRefund.js
@@ -2,26 +2,9 @@
 
 import { Far } from '@agoric/marshal';
 
-
-/**
- * This is a very trivial contract to explain and test Zoe.
- * AutomaticRefund just gives you back what you put in.
- * AutomaticRefund tells Zoe to complete the
- * offer, which gives the user their payout through Zoe. Other
- * contracts will use these same steps, but they will have more
- * sophisticated logic and interfaces.
- *
- * Since the contract doesn't attempt any reallocation, the offer can contain
- * anything in `give` and `want`. The amount in `give` will be returned, and
- * `want` will be ignored.
- *
- * @type {ContractStartFn}
- * @param {ContractFacet} zcf
- */
 const start = zcf => {
   let offersCount = 0n;
 
-  /** @type {OfferHandler} */
   const refund = seat => {
     offersCount += 1n;
     seat.exit();
@@ -29,7 +12,6 @@ const start = zcf => {
   };
   const makeRefundInvitation = () => zcf.makeInvitation(refund, 'getRefund');
 
-  /** @type {AutomaticRefundPublicFacet} */
   const publicFacet = Far('publicFacet', {
     getOffersCount: () => offersCount,
     makeInvitation: makeRefundInvitation,
