--- packages/zoe/test/bundle-source.js
@@ -1,4 +1,0 @@
-import bundleSource from '@agoric/bundle-source';
-
-// Use the new format.
-export default src => bundleSource(src, 'nestedEvaluate');
